initial commit of branch "single"
