--- pass-doi-service-core/src/docker/Dockerfile
@@ -1,2 +1,0 @@
-FROM jetty:9.4-jre8
-ADD target/pass-doi-service-core-0.0.1-SNAPSHOT.war /var/lib/jetty/webapps
