--- pass-doi-service-core/src/main/webapp/WEB-INF/web.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="3.0" xmlns="http://java.sun.com/xml/ns/javaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
-    <display-name>PASS DOI Service</display-name>
-    <servlet>
-        <servlet-name>doi-service</servlet-name>
-        <servlet-class>pass.doi.service.core.PassDoiServlet</servlet-class>
-        <load-on-startup>1</load-on-startup>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>doi-servlet</servlet-name>
-        <url-pattern>/doiServlet</url-pattern>
-    </servlet-mapping>
-
-</web-app>
-
-
