+++ src/test/resources/esconfig-3.4.json
